storage/spanlatch: rename storage/cmdq to storage/spanlatch
This change renames `storage/cmdq` to `storage/spanlatch`. The package
will house the new `spanlatch.Manager` type, which will handle the
acquisition and release of span-latches.

This works off of the definition for latches presented by Goetz Graefe
in https://15721.courses.cs.cmu.edu/spring2016/papers/a16-graefe.pdf
(see https://i.stack.imgur.com/fSRzd.png).

The files are not changes in this commit.

Release note: None
